🚚 Move: all TypeScript files to `src/lib/scripts/`
