+++ src/lib/components/FadeInUpAnim.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
 	import { onMount } from 'svelte';
 	import { browser } from '$app/environment';
-	import { addClassOnVisible } from '$lib/util';
+	import { addClassOnVisible } from '$lib/scripts/util';
 
 	let container: HTMLElement;
