+++ src/lib/scripts/easing.ts
