+++ src/lib/scripts/i18n/locales/ja.json
