+++ src/lib/scripts/util.ts
@@ -9,7 +9,7 @@
  * <script lang="ts">
  *     import { onMount } from 'svelte';
  *     import { browser } from '$app/environment';
- *     import { addClassOnVisible } from '$lib/util';
+ *     import { addClassOnVisible } from '$lib/scripts/util';
  *
  *     onMount(fadeIn);
  *     if (browser) window.addEventListener('scroll', fadeIn);
@@ -103,7 +103,7 @@ export function toggleScrollPrevention(prevent: boolean) {
  *
  * ```svelte
  * <script lang="ts">
- *     import { parallaxStyle } from '$lib/util';
+ *     import { parallaxStyle } from '$lib/scripts/util';
  *
  * 	   let scrollY: number;
  *	   $: parallax = parallaxStyle(scrollY);
