+++ src/lib/scripts/variables.ts
