+++ src/routes/+page.ts
@@ -1,5 +1,5 @@
 import type { PageLoad } from './$types';
-import type { ArticleMetadata } from '$lib/types';
+import type { ArticleMetadata } from '$lib/scripts/types';
 
 export const load: PageLoad = async ({ fetch }): Promise<{ articles: ArticleMetadata[] }> => {
 	const ARTICLE_LIMIT = 3;
