+++ src/routes/api/articles/thumbnail-imgs/+server.ts
@@ -1,4 +1,4 @@
-import { fetchThumbnailImgFmt } from '$lib/fetchers';
+import { fetchThumbnailImgFmt } from '$lib/scripts/fetchers';
 import { json } from '@sveltejs/kit';
 
 export const prerender = true;
