+++ src/routes/blog/articles/[slug]/+page.svelte
@@ -8,8 +8,8 @@
 
 	import type { PageData } from './$types';
 	import { page } from '$app/stores';
-	import { SITE_URL } from '$lib/variables';
-	import { parallaxStyle, idToDate } from '$lib/util';
+	import { SITE_URL } from '$lib/scripts/variables';
+	import { parallaxStyle, idToDate } from '$lib/scripts/util';
 	import { _ } from 'svelte-i18n';
 	import { cubicOut } from 'svelte/easing';
 	import { scale } from 'svelte/transition';
