+++ src/routes/feed/+server.ts
@@ -1,8 +1,8 @@
 import type { RequestHandler } from './$types';
-import { fetchArticles, fetchTags } from '$lib/fetchers';
-import { SITE_URL, COPYRIGHT, BLOG_NAME, CONTACT_EMAIL_ADDRESS } from '$lib/variables';
+import { fetchArticles, fetchTags } from '$lib/scripts/fetchers';
+import { SITE_URL, COPYRIGHT, BLOG_NAME, CONTACT_EMAIL_ADDRESS } from '$lib/scripts/variables';
 import { v5 as uuidv5 } from 'uuid';
-import { idToDate } from '$lib/util';
+import { idToDate } from '$lib/scripts/util';
 
 export const prerender = true;
