+++ src/routes/sitemap.xml/+server.ts
@@ -1,6 +1,6 @@
 import type { RequestHandler } from './$types';
-import { fetchArticles } from '$lib/fetchers';
-import { SITE_URL } from '$lib/variables';
+import { fetchArticles } from '$lib/scripts/fetchers';
+import { SITE_URL } from '$lib/scripts/variables';
 
 export const prerender = true;
